pyproject.toml
[project]
name = "nowcasting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.0",
    "langchain>=1.1.0",
    "langgraph>=1.0.4",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pillow>=10.0.0",
]
